refactor(node): rename processors to resource
