+++ opentron/src/manager/mod.rs
@@ -17,7 +17,7 @@ pub mod actuators;
 pub mod controllers;
 pub mod executor;
 pub mod governance;
-pub mod processors;
+pub mod resource;
 
 type Error = Box<dyn ::std::error::Error>;
 type Result<T, E = Error> = ::std::result::Result<T, E>;
